+++ lib/membrane/bin/action.ex
@@ -15,7 +15,8 @@ defmodule Membrane.Bin.Action do
 
   By default, component setup ends with the end of `c:Membrane.Bin.handle_setup/2` callback.
   If `{:setup, :incomplete}` is returned there, setup lasts until `{:setup, :complete}`
-  is returned from antoher callback.
+  is returned from antoher callback. You cannot return `{:setup, :incomplete}` from any other
+  callback than `c:Membrane.Bin.handle_setup/2`.
 
   Untils the setup lasts, the component won't enter `:playing` playback.
   """